Move tooltip test up to PresenceIconView
